Moved server.js to app.js
