+++ FrugalApp/package.json
@@ -1,6 +1,5 @@
 {
   "name": "frugalapp",
-  "main": "server.js",
   "version": "0.0.0",
   "private": true,
   "scripts": {
